Keep Xamarin.Forms dependency low... https://github.com/xamarin/Xamarin.Forms/issues/5652#issuecomment-498163046 Now at Xamarin.Forms 3.3

+++ CSharpMath.Forms.Example/CSharpMath.Forms.Example/EditorPage.xaml.cs
@@ -29,7 +29,7 @@ namespace CSharpMath.Forms.Example {
           case nameof(Entry.Text):
             var input = entry.Text.Split(' ')[1];
             foreach (var c in input)
-              keyboard.KeyPress((Editor.MathKeyboardInput)c);
+              keyboard.KeyPress((Editor.MathKeyboardInput)(int)c);
             entry.Text = Text;
             // Setting text also sets the cursor to the end
             entry.CursorPosition = CursorPosition;
